pyproject.toml
[tool.poetry]
name = "thoughtful-investor"
version = "0.1.0"
description = "A telegram bot holding TOMO"
authors = ["etienne-napoleone <etienne.napoleone@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
colorlog = "^4.0"
python-telegram-bot = "^11.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.7"

[tool.poetry.scripts]
thoughtful-investor = "thoughtful_investor.main:entrypoint"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
